Conformity: only test code with non-module Jar.
The conformity testing was using the module name to limit what types where tested. For jars that do not run under JPMS this meant types inherited from dependencies were also being tested.  This is inefficient and incorrect.  The tests now use the jar file name. This requires a few setup changes, which are documented.